// src/interface/arDn.d.ts
export interface AR_DN {
    DOCKEY: number;
    DOCNO: string;
    DOCNOEX?: string;
    CODE: string;
    DOCAMT: number;
    PAYMENTAMT: number;
    DESCRIPTION: string;
    DOCDATE: string;
    PROJECT: string;
    DUEDATE: string;
    TERMS: string;
    CANCELLED: string;
    LOCALDOCAMT: number;
}